+++ packages/core/src/index.ts
@@ -25,6 +25,7 @@ export * from "./components/heading/Heading";
 export * from "./types/ElementProps";
 export * from "./types/DeepPartial";
 export * from "./types/Omit";
+export * from "./types/PickByValue";
 
 export * from "./hooks/UseArraySet";
 export * from "./hooks/UseBoolean";
@@ -44,5 +45,6 @@ export * from "./hooks/UseForwardedRef";
 export * from "./hooks/UseTimeoutState";
 
 export * from "./utils/SwitchCaseExhauster";
+export * from "./utils/TruthyKeysAsList";
 export * from "./utils/parsers/NumberParser";
 export { booleanOrNumberToNumber } from "./utils/BooleanOrNumberToNumber";