+++ packages/core/src/types/PickByValue.ts
@@ -1,0 +1,4 @@
+export type PickByValue<T, ValueType> = Pick<
+  T,
+  { [Key in keyof T]-?: T[Key] extends ValueType ? Key : never }[keyof T]
+>;