+++ packages/filter/src/features/search-filter/components/SearchFilterDrawer.tsx
@@ -1,32 +1,26 @@
-import {
-  SearchFilterAction,
-  SearchFilterActions,
-  SearchFilterState,
-} from "../redux/SearchFilterRedux";
 import * as React from "react";
-import { Dispatch, PropsWithChildren, useCallback } from "react";
+import { useCallback } from "react";
 import { Drawer, DrawerProps } from "@stenajs-webui/modal";
 import { cssColor } from "@stenajs-webui/theme";
 import { Column } from "@stenajs-webui/core";
 import { SearchFilterPanelHeader } from "./SearchFilterPanelHeader";
+import { useSearchFilterState } from "../context/SearchFilterStateContext";
+import { useSearchFilterDispatch } from "../context/SearchFilterDispatchContext";
+import { useSearchFilterActions } from "../context/SearchFilterActionsContext";
 
-interface SearchFilterDrawerProps<TFormModel, TSectionKey extends string>
-  extends Omit<DrawerProps, "isOpen" | "onRequestClose"> {
-  state: SearchFilterState<TFormModel>;
-  dispatch: Dispatch<SearchFilterAction<TFormModel>>;
-  actions: SearchFilterActions<TFormModel, TSectionKey>;
-}
-export const SearchFilterDrawer = function SearchFilterDrawer<
-  TFormModel,
-  TSectionKey extends string
->({
-  state,
-  dispatch,
-  actions,
+interface SearchFilterDrawerProps
+  extends Omit<DrawerProps, "isOpen" | "onRequestClose"> {}
+
+export const SearchFilterDrawer: React.FC<SearchFilterDrawerProps> = ({
   children,
   ...drawerProps
-}: PropsWithChildren<SearchFilterDrawerProps<TFormModel, TSectionKey>>) {
-  const { open } = state.settings;
+}) => {
+  const {
+    settings: { open },
+  } = useSearchFilterState();
+  const dispatch = useSearchFilterDispatch();
+  const actions = useSearchFilterActions();
+
   const closeDrawer = useCallback(() => {
     dispatch(actions.closeFilters());
   }, [actions, dispatch]);