+++ packages/filter/src/features/search-filter/components/SearchFilterSection.tsx
@@ -1,23 +1,15 @@
 import * as React from "react";
-import { Dispatch, PropsWithChildren, useCallback } from "react";
+import { PropsWithChildren, useCallback } from "react";
 import { Row } from "@stenajs-webui/core";
-import {
-  SearchFilterAction,
-  SearchFilterActions,
-  SearchFilterState,
-} from "../redux/SearchFilterRedux";
 import { Collapsible, CollapsibleContent } from "@stenajs-webui/panels";
 import { lowerCase, upperFirst } from "lodash";
 import { Banner, FlatButton, Spinner } from "@stenajs-webui/elements";
+import { useSearchFilterState } from "../context/SearchFilterStateContext";
+import { useSearchFilterDispatch } from "../context/SearchFilterDispatchContext";
+import { useSearchFilterActions } from "../context/SearchFilterActionsContext";
 
-export interface SearchFilterSectionProps<
-  TFormModel,
-  TSectionKey extends string
-> {
+export interface SearchFilterSectionProps<TSectionKey extends string> {
   sectionId: TSectionKey;
-  state: SearchFilterState<TFormModel>;
-  actions: SearchFilterActions<TFormModel, TSectionKey>;
-  dispatch: Dispatch<SearchFilterAction<TFormModel>>;
   contentLeft?: React.ReactNode;
   contentRight?: React.ReactNode;
   disableContentPadding?: boolean;
@@ -28,22 +20,22 @@ export interface SearchFilterSectionProps<
 }
 
 export const SearchFilterSection = function SearchFilterSection<
-  TFormModel,
   TSectionKey extends string
 >({
   sectionId,
   label,
   contentLeft,
   contentRight,
-  state,
-  dispatch,
-  actions,
   loading,
   error,
   onRetry,
   disableContentPadding,
   children,
-}: PropsWithChildren<SearchFilterSectionProps<TFormModel, TSectionKey>>) {
+}: PropsWithChildren<SearchFilterSectionProps<TSectionKey>>) {
+  const state = useSearchFilterState();
+  const dispatch = useSearchFilterDispatch();
+  const actions = useSearchFilterActions();
+
   const expanded = state.expandedSections.values[sectionId] ?? false;
 
   const onClickLabel = useCallback(() => {