+++ packages/filter/src/features/search-filter/context/SearchFilterModelContext.ts
@@ -1,0 +1,11 @@
+import { createContext, useContext } from "react";
+
+export const SearchFilterModelContext = createContext<unknown>(null as unknown);
+
+export const useSearchFilterModel = <TFormModel>(): TFormModel => {
+  const c = useContext(SearchFilterModelContext);
+  if (!c) {
+    throw new Error("Missing search filter state context.");
+  }
+  return c as TFormModel;
+};