+++ packages/filter/src/features/search-filter/features/chips/SearchFilterChips.tsx
@@ -1,25 +1,22 @@
 import * as React from "react";
-import { Dispatch, useCallback } from "react";
-import {
-  SearchFilterAction,
-  SearchFilterActions,
-} from "../../redux/SearchFilterRedux";
+import { useCallback } from "react";
 import { Row } from "@stenajs-webui/core";
 import { ChipSpacer } from "./ChipSpacer";
 import { SecondaryButton } from "@stenajs-webui/elements";
+import { useSearchFilterDispatch } from "../../context/SearchFilterDispatchContext";
+import { useSearchFilterActions } from "../../context/SearchFilterActionsContext";
 
 interface SearchFilterChipsProps {
-  dispatch: Dispatch<SearchFilterAction<any>>;
-  actions: SearchFilterActions<any, any>;
   disableClearAllButton?: boolean;
 }
 
 export const SearchFilterChips: React.FC<SearchFilterChipsProps> = ({
   children,
-  actions,
-  dispatch,
   disableClearAllButton = false,
 }) => {
+  const dispatch = useSearchFilterDispatch();
+  const actions = useSearchFilterActions();
+
   const onClickClearAll = useCallback(
     () => dispatch(actions.clearFormModel()),
     [actions, dispatch]