+++ packages/filter/src/features/search-filter/section-factories/boolean-record/BooleanRecordTypes.ts
@@ -1,4 +1,5 @@
 export type BooleanRecord = Record<string, boolean>;
+
 export interface BooleanRecordOption {
   value: string;
   label: string;