+++ packages/filter/src/features/search-filter/section-factories/boolean-record/components/SimpleCheckboxListSection.tsx
@@ -3,33 +3,27 @@ import {
   SearchFilterSection,
   SearchFilterSectionProps,
 } from "../../../components/SearchFilterSection";
-import { Column } from "@stenajs-webui/core";
+import { Column, Text } from "@stenajs-webui/core";
 import { CollapsibleClickableContent } from "@stenajs-webui/panels";
 import { Checkbox, ValueAndOnValueChangeProps } from "@stenajs-webui/forms";
 import { BooleanRecord, BooleanRecordOptions } from "../BooleanRecordTypes";
 
-export interface CheckboxSectionProps<TFormModel, TSectionKey extends string>
-  extends SearchFilterSectionProps<TFormModel, TSectionKey>,
+export interface CheckboxSectionProps<TSectionKey extends string>
+  extends SearchFilterSectionProps<TSectionKey>,
     ValueAndOnValueChangeProps<BooleanRecord> {
   options?: BooleanRecordOptions;
 }
 
-export const CheckboxSection = <TFormModel, TSectionKey extends string>({
-  actions,
+export const SimpleCheckboxListSection = <TSectionKey extends string>({
   options,
-  dispatch,
   error,
   loading,
   sectionId,
-  state,
   value,
   onValueChange,
-}: CheckboxSectionProps<TFormModel, TSectionKey>): React.ReactElement => (
+}: CheckboxSectionProps<TSectionKey>): React.ReactElement => (
   <SearchFilterSection
     sectionId={sectionId}
-    state={state}
-    actions={actions}
-    dispatch={dispatch}
     loading={loading}
     error={error}
     disableContentPadding
@@ -49,7 +43,7 @@ export const CheckboxSection = <TFormModel, TSectionKey extends string>({
           }}
           contentLeft={<Checkbox tabIndex={-1} value={value?.[d.value]} />}
         >
-          {d.label}
+          <Text>{d.label}</Text>
         </CollapsibleClickableContent>
       ))}
     </Column>