+++ CHANGELOG.md
@@ -24,6 +24,13 @@ Please see stories for an example.
 - Added `ref` prop.
 - Added `display` prop so that you can e.g. display it inline.
 
+### New utils in `core`
+
+- `truthyKeysAsList`, returns all keys that have a truthy value.
+- `PickByValue`, a type that returns a new type with picked fields that match a type.
+
+`type StringFieldsOnUser = PickByValue<User, string>;`
+
 ### Breaking changes
 
 - Icon no longer has props `hoverIcon` and `hoverColor`.